requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
asyncmy==0.2.11
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.116.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.0.0
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.4.0
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.32
starlette==0.46.2
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0